+++ CHANGELOG.md
@@ -1,3 +1,17 @@
-## 0.0.1
+## 0.9.2-nullsafe
 
-* TODO: Describe initial release.
+Fix some problems that can't get the proxy configuration correctly under Windows
+
+## 0.9.1-nullsafe
+
+Fix problems with pubspecs on mac os and iOS
+
+## 0.9.0-nullsafe
+
+Pre-release with support for:
+
+* Android
+* iOS
+* Windows
+* MacOS
+* Nullsafety